requirements.txt
requests
python-dotenv
pyautogui
pytesseract
Pillow
pyttsx3
SpeechRecognition
playwright
PyAudio
